+++ write-service/src/service/UserService.js
@@ -1,5 +1,6 @@
 import { logger } from '../config/winston-logger.js'
 import { UserRepository } from '../repositories/UserRepository.js'
+import { DuplicateError } from '../util/Errors/DuplicateError.js'
 import { convertMongoCreateAtToISOdate } from '../util/index.js'
 import { validateNotUndefined } from '../util/validate.js'
 import { MessageBroker } from './MessageBroker.js'
@@ -36,11 +37,14 @@ export class UserService {
       await this.broker.sendMessage(process.env.USER_REGISTER_TOPIC, JSON.stringify(createdData))
       return createdData
     } catch (e) {
-      try {
-        logger.info('Succesfully cleaned up Kafka registration...')
-        await this.userRepo.deleteOneRecord({ userId: registrationData?.userId })
-      } catch (e) {
-        logger.error('Failed to cleanup Kafka registration...')
+      if (e instanceof DuplicateError === false) {
+        try {
+          logger.info('Succesfully cleaned up Kafka registration...')
+          await this.userRepo.deleteOneRecord({ userId: registrationData?.userId })
+        } catch (e) {
+          logger.error('Failed to cleanup Kafka registration...')
+          throw e
+        }
       }
       throw e
     }
